Added DocumentTagSchemaEvents

+++ lambda-api/src/main/java/com/formkiq/stacks/api/handler/DocumentTagValueRequestHandler.java
@@ -24,8 +24,11 @@
 package com.formkiq.stacks.api.handler;
 
 import static com.formkiq.aws.services.lambda.ApiResponseStatus.SC_OK;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.Map;
 import com.amazonaws.services.lambda.runtime.LambdaLogger;
+import com.formkiq.aws.dynamodb.model.DocumentItem;
 import com.formkiq.aws.services.lambda.ApiAuthorizer;
 import com.formkiq.aws.services.lambda.ApiGatewayRequestEvent;
 import com.formkiq.aws.services.lambda.ApiGatewayRequestEventUtil;
@@ -34,7 +37,9 @@ import com.formkiq.aws.services.lambda.ApiMessageResponse;
 import com.formkiq.aws.services.lambda.ApiRequestHandlerResponse;
 import com.formkiq.aws.services.lambda.ApiResponse;
 import com.formkiq.aws.services.lambda.AwsServiceCache;
-import com.formkiq.aws.services.lambda.NotFoundException;
+import com.formkiq.aws.services.lambda.ValidationError;
+import com.formkiq.aws.services.lambda.exceptions.NotFoundException;
+import com.formkiq.aws.services.lambda.exceptions.ValidationException;
 import com.formkiq.stacks.api.CoreAwsServiceCache;
 import com.formkiq.stacks.dynamodb.DocumentService;
 
@@ -62,6 +67,17 @@ public class DocumentTagValueRequestHandler
     CoreAwsServiceCache cacheService = CoreAwsServiceCache.cast(awsservice);
     DocumentService documentService = cacheService.documentService();
 
+    DocumentItem item = documentService.findDocument(siteId, documentId);
+    if (item == null) {
+      throw new NotFoundException("Document " + documentId + " not found.");
+    }
+    
+    Collection<ValidationError> errors =
+        awsservice.documentTagSchemaEvents().removeTagsEvent(siteId, item, Arrays.asList(tagKey));
+    if (!errors.isEmpty()) {
+      throw new ValidationException(errors);
+    }
+    
     boolean removed = documentService.removeTag(siteId, documentId, tagKey, tagValue);
     if (!removed) {
       throw new NotFoundException("Tag/Value combination not found.");
